Furter refactoring into cogs. Helper & Toy cogs created. Extracted channel decorator & DM check into predicates.py
