+++ stGermain.py
@@ -2,10 +2,11 @@ import discord
 import logging
 from discord.ext import commands
 
-from src.cogs.tarotCog import Tarot
+from src.cogs.tarot import Tarot
+from src.cogs.helper import Helper
+from src.cogs.toys import Toy
 from src.server import keepAlive, token
-from src.magicEight.magicEight import magicEightBall
-from src.guidance.userGuide import userGuide
+from src.backend import predicates
 from src.recipes.recipe import respondForRecipe
 
 # Bot setup.
@@ -20,27 +21,14 @@ bot = commands.Bot(
     intents=intents
 )
 bot.add_cog(Tarot(bot))
+bot.add_cog(Helper(bot))
+bot.add_cog(Toy(bot))
 
 # Get correct API token path.
 TOKEN = token.replOrLocal(token.repl, token.devFlag)
 
 # Channel ID for inChannel check.
 whiteLodgeChannel = 817823496352169985
-
-
-# Decorator for limiting commands per channel.
-def inChannels(*args):
-    def predicate(ctx):
-        return ctx.message.channel.id in args
-
-    return commands.check(predicate)
-
-
-async def checkDM(ctx):
-    if ctx.author.bot:
-        return
-    if isinstance(ctx.channel, discord.channel.DMChannel):
-        await ctx.author.send("DM")
 
 
 @bot.event
@@ -57,29 +45,13 @@ async def on_ready():
 
 
 @bot.command()
-async def guidance(ctx):
-    await ctx.send(userGuide)
-
-
-@bot.command()
-async def add(ctx, left: int, right: int):
-    # Adds two numbers
-    await ctx.send(left + right)
-
-
-@bot.command()
-async def magicEight(ctx, *, message='question'):
-    await magicEightBall(ctx, message)
-
-
-@bot.command()
 async def createRecipe(ctx, *, message=''):
     await respondForRecipe(ctx, message)
 
 
 @bot.listen()
 async def on_message(ctx):
-    await checkDM(ctx)
+    await predicates.checkDM(ctx)
 
 
 if token.repl is True:
